Statistics STF => applyStatsStf function
✅ 3/3 tests passed (in tiny and full).

The function is sufficient to update the state based on extrinsics and epoch changes to pi.current and pi.last, and i believe that is all that is required.

Tau in the input doesnt update the post state in the test vectors , because it is handled elsewhere, i dont know exactly why it was used in these vectors, same with kappa_prime, they are not manipulated in any way. Again unless i am missing something it is dummy data.